--- Task6/config/config.yaml
@@ -1,14 +1,0 @@
-comic:
-  source_url : https://xkcd.com
-  parallel : 7
-
-server:
-  port : 8080
-
-db:
-  db_file : "database/database.json"
-  dsn: "internal/database/sqlite_DB/xkcdDB.db"
-
-
-# Количество логических ядер на моём ноутбуке равно 8. 
-# Сделала -1 на всякий случай, чтобы одно ядро было свободно для других задач.
