--- Task6/internal/database/sqlite_db/000001_keywords.up.sql
@@ -1,10 +1,0 @@
-CREATE TABLE IF NOT EXISTS keyword (
-    number INTEGER PRIMARY KEY,
-    url TEXT,
-    keywords TEXT
-);
-
-CREATE TABLE IF NOT EXISTS words_index (
-    word TEXT PRIMARY KEY,
-    numbers TEXT
-);
