+++ Task7/config/config.yaml
@@ -1,0 +1,17 @@
+comic:
+  source_url : https://xkcd.com
+  parallel : 7
+
+server:
+  port : 8080
+  concurrency_limit : 5
+  rate_limit : 10
+  rate_limit_per_second : 2
+
+db:
+  db_file : "database/database.json"
+  dsn: "pkg/repository/database/xkcdDB.db"
+
+
+# Количество логических ядер на моём ноутбуке равно 8. 
+# Сделала -1 на всякий случай, чтобы одно ядро было свободно для других задач.
