+++ Task7/go.mod
@@ -1,0 +1,29 @@
+module task7
+
+go 1.22.1
+
+require (
+	github.com/ilyakaznacheev/cleanenv v1.5.0
+	github.com/kljensen/snowball v0.9.0
+	github.com/mattn/go-sqlite3 v1.14.17
+	golang.org/x/sync v0.7.0
+)
+
+require (
+	github.com/hashicorp/errwrap v1.1.0 // indirect
+	github.com/hashicorp/go-multierror v1.1.1 // indirect
+	github.com/kr/pretty v0.3.1 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	go.uber.org/atomic v1.11.0 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
+)
+
+require (
+	github.com/BurntSushi/toml v1.3.2 // indirect
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/golang-migrate/migrate/v4 v4.17.1
+	github.com/joho/godotenv v1.5.1 // indirect
+	golang.org/x/time v0.5.0
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+	olympos.io/encoding/edn v0.0.0-20201019073823-d3554ca0b0a3 // indirect
+)
