+++ Task7/pkg/handler/handler.go
@@ -1,0 +1,22 @@
+package handler
+
+import (
+	"net/http"
+	"task7/pkg/service"
+)
+
+type Handler struct {
+	services *service.Service
+}
+
+func NewHandler(services *service.Service) *Handler {
+	return &Handler{services: services}
+}
+
+func (h *Handler) Init() {
+
+	//все три хендлера пропускаем через функцию - ограничитель запросов от одного клиента
+	http.HandleFunc("GET /pics", perClientRateLimiter(http.HandlerFunc(h.handlePics)))
+	http.HandleFunc("POST /update", perClientRateLimiter(http.HandlerFunc(h.handleUpdate)))
+	http.HandleFunc("POST /login", perClientRateLimiter(http.HandlerFunc(h.handleLogin)))
+}
