+++ Task7/schema/000001_keywords.down.sql
