+++ Task7/written_comics.txt
@@ -1,0 +1,1 @@
+&[]
